+++ src/err.rs
@@ -1,3 +1,6 @@
+use crate::encoding::RecodeCause;
+use crate::{encoding, Encoding, Str};
+
 /// Error encountered while re-encoding a [`Str`](crate::Str) or [`CStr`](crate::CStr) into another
 /// format
 #[derive(Clone, Debug, PartialEq)]
@@ -25,5 +28,40 @@ impl RecodeError {
     /// character (though recoding may fail again immediately due to another invalid character).
     pub fn char_len(&self) -> usize {
         self.char_len as usize
+    }
+}
+
+/// Error encountered while re-encoding a [`Str`](crate::Str) or [`CStr`](crate::CStr) into another
+/// format in a pre-allocated buffer
+pub struct RecodeIntoError<'a, E: Encoding> {
+    pub(crate) input_used: usize,
+    pub(crate) str: &'a Str<E>,
+    pub(crate) cause: RecodeCause,
+}
+
+impl<'a, E: Encoding> RecodeIntoError<'a, E> {
+    pub(crate) fn from_recode(err: encoding::RecodeError, str: &'a Str<E>) -> Self {
+        RecodeIntoError {
+            input_used: err.input_used(),
+            str,
+            cause: err.cause().clone(),
+        }
+    }
+
+    /// The length of valid data in the input before the error was encountered. Calling
+    /// [`recode_into`](crate::Str::recode_into) again on the input sliced down to this length will succeed.
+    pub fn valid_up_to(&self) -> usize {
+        self.input_used
+    }
+
+    /// The portion of the buffer with valid data written into it, as a [`Str`] in the desired
+    /// encoding.
+    pub fn output_valid(&self) -> &'a Str<E> {
+        self.str
+    }
+
+    /// The reason encoding stopped. See [`RecodeCause`] for more details on possible reasons.
+    pub fn cause(&self) -> &RecodeCause {
+        &self.cause
     }
 }