+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# Unreleased
+
+### Added
+
+- `Str::recode_into` - non-allocating variant of recode, returns the `Str` written
+
 # [0.1.2] - 2024-08-08
 
 ### Added